feature_flags: Move feature flag definitions to their own module
